# spring/api-docs.yaml
openapi: 3.0.1
info:
  title: Конспект по языку Java
  description: Конспект по языку Java на русском языке
  contact:
    name: Токовенко Виктор
    url: https://github.com/Lorgar-Avrelian?tab=repositories
    email: victor-14-244@mail.ru
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /books:
    get:
      tags:
        - 2 Книги
        - Поиск
      summary: Список книг
      description: Вывести список всех доступных книг
      operationId: getAllBooks
      parameters:
        - name: authorOrTitle
          in: query
          description: Часть ФИО автора или названия книги
          required: false
          schema:
            type: string
          example: Пушкин
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    put:
      tags:
        - 2 Книги
        - Книга
      summary: Редактировать
      description: Отредактировать информацию о книге
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "400":
          description: Bad Request
      security:
        - basicAuth: []
    post:
      tags:
        - 2 Книги
        - Книга
      summary: Создать
      description: Добавить информацию о книге
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBookDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "405":
          description: Method Not Allowed
      security:
        - basicAuth: []
...
  /books/{id}/cover:
    get:
      tags:
        - 2 Книги
        - Изображения
      summary: Посмотреть
      description: Посмотреть обложку для книги
      operationId: downloadCover
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            multipart/form-data:
              schema:
                $ref: "#/components/schemas/BookCover"
        "404":
          description: Not Found
        "502":
          description: Bad Gateway
      security:
        - basicAuth: []
    post:
      tags:
        - 2 Книги
        - Изображения
      summary: Загрузить
      description: Загрузить обложку для книги
      operationId: uploadCover
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: Изображение обложки книги
                  format: binary
      responses:
        "200":
          description: OK
          content:
            multipart/form-data:
              schema:
                $ref: "#/components/schemas/BookCover"
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      security:
        - basicAuth: []
...
  /books/{id}:
    get:
      tags:
        - 2 Книги
        - Книга
      summary: Найти
      description: Найти информацию о книге по ID
      operationId: readBook
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "404":
          description: Not Found
      security:
        - basicAuth: []
    delete:
      tags:
        - 2 Книги
        - Книга
      summary: Удалить
      description: Удалить информацию о книге по ID
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDTO"
        "403":
          description: Forbidden
      security:
        - basicAuth: []
  /books/{id}/cover/preview:
    get:
      tags:
        - 2 Книги
        - Изображения
      summary: Посмотреть превью
      description: Посмотреть превью обложки для книги
      operationId: downloadCover_1
      parameters:
        - name: id
          in: path
          description: ID книги в имеющемся списке книг
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: OK
          content:
            multipart/form-data:
              schema:
                type: array
                items:
                  type: string
                  format: byte
        "403":
          description: Forbidden
      security:
        - basicAuth: []
...
components:
  schemas:
    ReaderNoBooksDTO:
      title: Читатель
      maximum: 9223372036854775807
      minimum: 1
      required:
        - id
        - name
        - personalNumber
        - surname
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID читателя
          format: int64
          default: 1
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность читателя
      default: null
    Reader:
      title: Читатель
      maximum: 9223372036854775807
      minimum: 1
      required:
        - id
        - name
        - personalNumber
        - surname
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID читателя
          format: int64
          default: 1
        name:
          title: Имя
          maxLength: 12
          minLength: 2
          type: string
          description: Имя читателя
          default: Иван
        secondName:
          title: Отчество
          maxLength: 16
          minLength: 2
          type: string
          description: Отчество читателя
          default: Иванович
        surname:
          title: Фамилия
          maxLength: 30
          minLength: 2
          type: string
          description: Фамилия читателя
          default: Иванов
        personalNumber:
          title: Номер билета
          maximum: 2147483648
          minimum: 1
          type: integer
          description: Номер читательского билета
          format: int32
          default: 1
      description: Сущность читателя
      default: null
    BookDTO:
      title: Книга
      required:
        - author
        - id
        - title
        - year
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID книги
          format: int64
          default: 1
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название книги
          default: Война и мир
        author:
          title: Автор
          maxLength: 24
          minLength: 8
          type: string
          description: Автор книги
          default: Л.Н. Толстой
        year:
          title: Год
          minimum: 1970
          type: integer
          description: Год публикации книги
          format: int32
          default: 1986
        reader:
          $ref: "#/components/schemas/ReaderNoBooksDTO"
      description: Сущность книги
...
    NewBookDTO:
      title: Новая книга
      required:
        - authorFIO
        - title
        - year
      type: object
      properties:
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название книги
          default: Война и мир
        authorFIO:
          title: Автор
          maxLength: 24
          minLength: 8
          type: string
          description: Автор книги
          default: Л.Н. Толстой
        year:
          title: Год
          minimum: 1970
          type: integer
          description: Год публикации книги
          format: int32
          default: 1986
      description: Сущность новой книги
    Book:
      title: Книга
      maximum: 9223372036854775807
      minimum: 1
      required:
        - author
        - id
        - title
        - year
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID книги
          format: int64
          default: 1
        title:
          title: Название
          maxLength: 30
          minLength: 3
          type: string
          description: Название книги
          default: Война и мир
        author:
          title: Автор
          maxLength: 24
          minLength: 8
          type: string
          description: Автор книги
          default: Л.Н. Толстой
        year:
          title: Год
          minimum: 1970
          type: integer
          description: Год публикации книги
          format: int32
          default: 1986
        reader:
          $ref: "#/components/schemas/Reader"
      description: Сущность книги
      default: null
    BookCover:
      title: Обложка книги
      required:
        - filePath
        - fileSize
        - id
        - imagePreview
        - mediaType
      type: object
      properties:
        id:
          title: ID
          maximum: 9223372036854775807
          minimum: 1
          type: integer
          description: ID обложки книги
          format: int64
          default: 1
        filePath:
          title: Путь
          type: string
          description: Путь сохранения обложки книги
        fileSize:
          title: Размер
          type: integer
          description: Размер сохранённой обложки книги
          format: int32
        mediaType:
          title: Тип
          type: string
          description: Тип сохранённой обложки книги
        imagePreview:
          title: Превью
          type: string
          description: Превью обложки книги
          format: byte
        book:
          $ref: "#/components/schemas/Book"
      description: Сущность обложки для книг
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic